# openapi.yaml
openapi: 3.0.3
info:
  title: SmartCloudOps AI API
  description: Enterprise AI-powered cloud operations platform with real-time monitoring,
    anomaly detection, and automated remediation
  version: 5.0.0
  contact:
    name: SmartCloudOps Team
    email: support@smartcloudops.ai
    url: https://smartcloudops.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
- url: http://localhost:5000
  description: Development server
- url: https://api.smartcloudops.ai
  description: Production server
security:
- bearerAuth: []
...
paths:
...
  /api/performance/anomaly/detect:
    post:
      summary: Detect anomalies
      description: Detect anomalies in system metrics using ML
      tags:
      - Anomaly Detection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnomalyDetectionRequest'
      responses:
        '200':
          description: Anomaly detection completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnomalyDetectionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/performance/cache/stats:
    get:
      summary: Get cache statistics
      description: Retrieve Redis cache performance statistics
      tags:
      - Performance
      responses:
        '200':
          description: Cache statistics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      hit_rate:
                        type: number
                        format: float
                      total_requests:
                        type: integer
                      hits:
                        type: integer
                      misses:
                        type: integer
        '401':
          $ref: '#/components/responses/UnauthorizedError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: An error occurred
        code:
          type: string
          example: ERR_001
      required:
      - status
      - message
    ValidationError:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: Validation failed
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    Metrics:
      type: object
      properties:
        cpu_usage_percent:
          type: number
          format: float
          example: 45.5
        memory_usage_percent:
          type: number
          format: float
          example: 62.3
        disk_usage_percent:
          type: number
          format: float
          example: 38.7
        load_avg_1min:
          type: number
          format: float
          example: 2.1
        network_bytes_sent_rate:
          type: number
          format: float
          example: 1024.5
        network_bytes_recv_rate:
          type: number
          format: float
          example: 2048.3
    AnomalyDetectionRequest:
      type: object
      properties:
        metrics:
          $ref: '#/components/schemas/Metrics'
        use_cache:
          type: boolean
          default: true
      required:
      - metrics
    AnomalyDetectionResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          type: object
          properties:
            is_anomaly:
              type: boolean
            anomaly_score:
              type: number
              format: float
            confidence:
              type: number
              format: float
            timestamp:
              type: string
              format: date-time
...
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'